Moved `Content` components to separate file
