+++ src/components/Content.tsx
@@ -1,16 +1,59 @@
-import React from 'react';
+import { Box, Center, Flex, Text } from '@chakra-ui/react';
+import { useContext, useState } from 'react';
+import CurrentContentContext, {
+  CurrentContentLiterals,
+} from '../contexts/CurrentContentContext';
+import SizeContext from '../contexts/SizeContext';
 
-type props = {
-  additionalClass?: string;
+/* Content components */
+const ContentDashboard: React.FC = () => {
+  return <Text>Dashboard: Sales numbers</Text>;
 };
 
-const Content: React.FC<props> = ({ additionalClass }) => {
+const ContentPOS: React.FC = () => {
+  return <Text>POS: Actual POS display</Text>;
+};
+
+const ContentTransactions: React.FC = () => {
+  return <Text>Transactions: List of previous transactions</Text>;
+};
+
+const ContentInventory: React.FC = () => {
+  return <Text>Inventory: Current stock</Text>;
+};
+
+let ContentMap: Record<CurrentContentLiterals, JSX.Element> = {
+  home: <ContentDashboard />,
+  pos: <ContentPOS />,
+  transactions: <ContentTransactions />,
+  inventory: <ContentInventory />,
+};
+
+const Content: React.FC = () => {
+  const sizes = useContext(SizeContext);
+  const [contentLabel, setContentLabel] = useState<CurrentContentLiterals>(
+    'home'
+  );
+
   return (
-    <div className={[additionalClass].join(' ')}>
-      <div></div>
-      <div></div>
-      <div></div>
-    </div>
+    <Box h={`calc(100% - ${sizes.navbar.height})`} p="1%">
+      {/* Content */}
+      <Flex
+        h="100%"
+        // borderRadius={`${sizes.contentRadius} ${sizes.contentRadius} 0 0`}
+        borderRadius={sizes.content.radius}
+        // m="1%"
+        // bgColor={randomizedCSSrgb()}
+        bgColor="white"
+      >
+        <Center w="100%">
+          {/* <Text>Content</Text> */}
+          <CurrentContentContext.Consumer>
+            {({ label, changeContent }) => ContentMap[label]}
+          </CurrentContentContext.Consumer>
+        </Center>
+      </Flex>
+    </Box>
   );
 };
